quiet compiler warnings using appropriate type casting.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@9293 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/util_funcs.c
@@ -638,7 +638,7 @@ header_simple_table(struct variable *vp, oid * name, size_t * length,
             newname[*length - 1] = name[*length - 1];
         }
     }
-    if ((max >= 0 && (newname[*length - 1] > max)) ||
+    if ((max >= 0 && ((int)newname[*length - 1] > max)) ||
                ( 0 == newname[*length - 1] )) {
         if (var_len)
             *var_len = 0;
